# build/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API para explorar el mundo de Disney. La idea de la misma es que permite conocer y modificar los personajes que lo componen y entender en qué películas estuvieron.",
    "version": "1.0.0",
    "title": "Disney API",
    "contact": {
      "email": "deckerurbano@gamil.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "disney-api-challenge.herokuapp.com",
  "basePath": "/",
  "tags": [
    {
      "name": "character",
      "description": ""
    },
    {
      "name": "film",
      "description": ""
    },
    {
      "name": "user",
      "description": ""
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/character": {
      "get": {
        "tags": [
          "character"
        ],
        "summary": "Listado de Personajes",
        "description": "El listado muestra imagen, nombre.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Objeto Personaje",
            "schema": {
              "$ref": "#/definitions/character"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "character"
        ],
        "summary": "Agrega un personaje",
        "description": "",
        "operationId": "addPersonaje",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Nombre actualizado del personaje",
            "required": false,
            "type": "string"
          },
          {
            "name": "picture",
            "in": "formData",
            "description": "Imagen actualizado del personaje",
            "required": false,
            "type": "string"
          },
          {
            "name": "age",
            "in": "formData",
            "description": "Edad actualizado del personaje",
            "required": false,
            "type": "string"
          },
          {
            "name": "weight",
            "in": "formData",
            "description": "Peso actualizado del personaje",
            "required": false,
            "type": "string"
          },
          {
            "name": "history",
            "in": "formData",
            "description": "Historia actualizado del personaje",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Se agrego el personaje"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/character/{id}": {
      "get": {
        "tags": [
          "character"
        ],
        "summary": "Detalle de Personaje",
        "description": "En el detalle deberán listarse todos los atributos del personaje, como así también sus películas o series relacionadas.",
        "operationId": "getCharacterById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID de personaje a retornar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/character"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "character"
        ],
        "summary": "Actualiza  personaje por id",
        "description": "",
        "operationId": "updateCharacterById",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID del personaje a actualizar",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Nombre actualizado del personaje",
            "required": false,
            "type": "string"
          },
          {
            "name": "picture",
            "in": "formData",
            "description": "Imagen actualizado del personaje",
            "required": false,
            "type": "string"
          },
          {
            "name": "age",
            "in": "formData",
            "description": "Edad actualizado del personaje",
            "required": false,
            "type": "string"
          },
          {
            "name": "weight",
            "in": "formData",
            "description": "Peso actualizado del personaje",
            "required": false,
            "type": "string"
          },
          {
            "name": "history",
            "in": "formData",
            "description": "Historia actualizado del personaje",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Se actualizo el personaje"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "character"
        ],
        "summary": "Borra un personaje",
        "description": "",
        "operationId": "deleteCharacterById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID del personaje a borrar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Se ha borrado el personaje"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/character/filter": {
      "get": {
        "tags": [
          "character"
        ],
        "summary": "Podes filtrar un personaje por edad, peso, peliculas/series relacionadas",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "edad",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "peso",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "peliculas",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Objeto Personaje Filtrado",
            "schema": {
              "$ref": "#/definitions/character"
            }
          },
          "404": {
            "description": "No hay Personaje que cumplan con ese filtro"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/character/search/{name}": {
      "get": {
        "tags": [
          "character"
        ],
        "summary": "Busca por nombre de personaje",
        "description": "",
        "operationId": "searchNameCharacter",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "path",
            "description": "nombre del personaje a retornar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/character"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/film": {
      "get": {
        "tags": [
          "film"
        ],
        "summary": "Listado de Peliculas/Series",
        "description": "El listado muestra imagen, nombre.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Objeto Personaje",
            "schema": {
              "$ref": "#/definitions/Films"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "film"
        ],
        "summary": "Agrega una Pelicula/Serie",
        "description": "",
        "operationId": "createFilm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "picture",
            "in": "formData",
            "description": "Descripción de la Pelicula/Serie",
            "required": false,
            "type": "string"
          },
          {
            "name": "title",
            "in": "formData",
            "description": "Titulo de la Pelicula/Serie",
            "required": false,
            "type": "string"
          },
          {
            "name": "creation_date",
            "in": "formData",
            "description": "Dia de creacion de la Pelicula/Serie",
            "required": false,
            "type": "string"
          },
          {
            "name": "stars",
            "in": "formData",
            "description": "Calificación de la Pelicula/Serie",
            "required": false,
            "type": "string"
          },
          {
            "name": "genre",
            "in": "formData",
            "description": "Genero de la Pelicula/Serie",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Se creo una Pelicula/Serie"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/film/{id}": {
      "get": {
        "tags": [
          "film"
        ],
        "summary": "Detalle de una Pelicula/Serie",
        "description": "Devolverá todos los campos de la película o serie junto a los personajes asociados a la misma",
        "operationId": "getFilmById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID de la Pelicula/Serie a retornar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Films"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "film"
        ],
        "summary": "Actualiza una Pelicula/Serie por id",
        "description": "",
        "operationId": "updateFilmById",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID del personaje a actualizar",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "picture",
            "in": "formData",
            "description": "Descripción actualizado de la Pelicula/Serie",
            "required": false,
            "type": "string"
          },
          {
            "name": "title",
            "in": "formData",
            "description": "Titulo actualizado de la Pelicula/Serie",
            "required": false,
            "type": "string"
          },
          {
            "name": "creation_date",
            "in": "formData",
            "description": "Dia de creacion actualizado de la Pelicula/Serie",
            "required": false,
            "type": "string"
          },
          {
            "name": "stars",
            "in": "formData",
            "description": "Calificación actualizada de la Pelicula/Serie",
            "required": false,
            "type": "string"
          },
          {
            "name": "genre",
            "in": "formData",
            "description": "Genero actualizado de la Pelicula/Serie",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Se actualizo una Pelicula/Serie"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "film"
        ],
        "summary": "Borra una Pelicula/Serie",
        "description": "",
        "operationId": "deleteFilmById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID del una Pelicula/Serie a borrar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Se borro una Pelicula/Serie"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/search/{title}": {
      "get": {
        "tags": [
          "film"
        ],
        "summary": "Busca por nombre de la pelicula",
        "description": "",
        "operationId": "searchTitle",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "title",
            "in": "path",
            "description": "Titulo de la Pelicula/Serie a retornar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Films"
            }
          },
          "404": {
            "description": "No hay peliculas o series con este titulo"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/genre/{genre}": {
      "get": {
        "tags": [
          "film"
        ],
        "summary": "Busca por genero de pelicula/serie",
        "operationId": "searchWithGenre",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "genre",
            "in": "path",
            "description": "Genero de la Pelicula/Serie a retornar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Films"
            }
          },
          "404": {
            "description": "No hay peliculas o series con este genero"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/order/{order}": {
      "get": {
        "tags": [
          "film"
        ],
        "summary": "Detalle de una Pelicula/Serie",
        "description": "",
        "operationId": "orderFilms",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "order",
            "in": "path",
            "description": "Forman en la que ordenarla Pelicula/Serie (ASC/DESC)",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Films"
            }
          },
          "400": {
            "description": "El orden que elegio es incorrecto use asc o desc"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Crear un usuario",
        "description": "",
        "operationId": "register",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "email del usuario",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Password del usuario",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Se creo el usuario"
          }
        }
      }
    },
    "/auth/login": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Iniciar sesion",
        "description": "",
        "operationId": "login",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "email del usuario",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Password del usuario",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "El token expira en 5 minutos"
              }
            }
          },
          "203": {
            "description": "Error en email y/o contraseña"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "user-token",
      "in": "header"
    }
  },
  "definitions": {
    "character": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "picture": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "age": {
          "type": "string"
        },
        "weight": {
          "type": "string"
        },
        "history": {
          "type": "string"
        }
      }
    },
    "Films": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "picture": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "creation_date": {
          "type": "string",
          "format": "date-time"
        },
        "stars": {
          "type": "string"
        },
        "genre": {
          "type": "string"
        }
      }
    },
    "character_film": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "film_id": {
          "type": "integer"
        },
        "character_id": {
          "type": "integer"
        }
      }
    }
  }
}